token sign and verify

+++ controller/userController.js
@@ -1,12 +1,23 @@
 const db = require("../models")
+const jwt = require('jsonwebtoken');
+require('dotenv').config()
+const crypto = require("../helpers/crypto")
+
 
 const findAll = (req, res)=>{
-  db.User.findAll()
-  .then(rows=>{
-    res.send(rows)
-  }).catch(err=>{
-    res.send(err)
-  })
+  let token = req.headers.token
+  let decoded = jwt.verify(token, process.env.TOKEN)
+  if(decoded.role == "admin"){
+    db.User.findAll()
+    .then(rows=>{
+      res.send(rows)
+    }).catch(err=>{
+      res.send(err)
+    })
+  }
+  else{
+    res.send("Token Salah")
+  }
 }
 
 const findById = (req, res)=>{
@@ -18,31 +29,45 @@ const findById = (req, res)=>{
   })
 }
 const create = (req, res)=>{
-  db.User.create({
-    username : req.body.username,
-    password : req.body.password,
-    createdAt : new Date(),
-    updatedAt : new Date()
-  })
-  .then(()=>{
-    res.send("Created")
-  }).catch(err=>{
-    res.send(err)
-  })
+  let token = req.headers.token
+  let decoded = jwt.verify(token, process.env.TOKEN)
+  if(decoded.role == "admin"){
+    db.User.create({
+      username : req.body.username,
+      password : req.body.password,
+      role : req.body.role,
+      createdAt : new Date(),
+      updatedAt : new Date()
+    })
+    .then(()=>{
+      res.send("Created")
+    }).catch(err=>{
+      res.send(err)
+    })
+  }else{
+    res.send("Token Salah")
+  }
 }
 
 const destroy = (req, res)=>{
-  db.User.destroy({where : {id : req.params.id}})
-  .then(()=>{
-    res.send("telah dihapus")
-  }).catch(err=>{
-    res.send(err)
-  })
+  let token = req.headers.token
+  let decoded = jwt.verify(token,  process.env.TOKEN)
+  if(decoded.role == "admin"){
+    db.User.destroy({where : {id : req.params.id}})
+    .then(()=>{
+      res.send("telah dihapus")
+    }).catch(err=>{
+      res.send(err)
+    })
+  }else{
+    res.send("Token Salah")
+  }
 }
 const update = (req, res)=>{
   db.User.update({
     username : req.body.username,
     password : req.body.password,
+    role : req.body.role,
     createdAt : new Date(),
     updatedAt : new Date()
   }, {where : {id : req.params.id}})
@@ -53,10 +78,44 @@ const update = (req, res)=>{
   })
 }
 
+const signup = (req, res)=>{
+  db.User.create({
+    username : req.body.username,
+    password : req.body.password,
+    role : req.body.role,
+    createdAt : new Date(),
+    updatedAt : new Date()
+  })
+  .then(()=>{
+    res.send("Register Success")
+  }).catch(err=>{
+    res.send(err)
+  })
+}
+
+const login = (req, res)=>{
+  db.User.findOne({where : {username: req.body.username}})
+  .then(row=>{
+    if(crypto.secretkey(req.body.password, row.secret) == row.password){
+      let token = {}
+       token.token = jwt.sign({
+        username:row.username,
+        role:row.role
+      }, process.env.TOKEN)
+      res.send(token)
+      // res.send("Anda Telah Masuk")
+    }else{
+      res.send("User Tidak Ditemukan")
+    }
+  })
+}
+
 module.exports = {
     findAll,
     findById,
     create,
     destroy,
-    update
+    update,
+    signup,
+    login
 }
